refactor: Editor.kt files decoupled and new toggle buttons
