+++ app/src/androidTest/java/com/serranoie/app/minus/presentation/ui/e2e/budget/BudgetPeriodSheetEditModeE2ETest.kt
@@ -11,11 +11,11 @@ import com.serranoie.app.minus.R
 import com.serranoie.app.minus.domain.model.BudgetPeriod
 import com.serranoie.app.minus.domain.model.BudgetSettings
 import com.serranoie.app.minus.domain.model.BudgetState
-import com.serranoie.app.minus.presentation.ui.editor.BUDGET_PERIOD_APPLY_BUTTON_TAG
-import com.serranoie.app.minus.presentation.ui.editor.BUDGET_PERIOD_BUDGET_INPUT_TAG
-import com.serranoie.app.minus.presentation.ui.editor.BUDGET_PERIOD_EDIT_BUTTON_TAG
-import com.serranoie.app.minus.presentation.ui.editor.BUDGET_PERIOD_SHEET_TAG
-import com.serranoie.app.minus.presentation.ui.editor.BudgetPeriodSheet
+import com.serranoie.app.minus.presentation.ui.editor.sheets.BUDGET_PERIOD_APPLY_BUTTON_TAG
+import com.serranoie.app.minus.presentation.ui.editor.sheets.BUDGET_PERIOD_BUDGET_INPUT_TAG
+import com.serranoie.app.minus.presentation.ui.editor.sheets.BUDGET_PERIOD_EDIT_BUTTON_TAG
+import com.serranoie.app.minus.presentation.ui.editor.sheets.BUDGET_PERIOD_SHEET_TAG
+import com.serranoie.app.minus.presentation.ui.editor.sheets.BudgetPeriodSheet
 import com.serranoie.app.minus.presentation.ui.theme.MinusTheme
 import org.junit.Rule
 import org.junit.Test
