+++ app/src/main/java/com/serranoie/app/minus/presentation/ui/editor/EditorTypes.kt
@@ -1,11 +1,5 @@
 package com.serranoie.app.minus.presentation.ui.editor
 
-/**
- * Animation state for editor transitions.
- */
 enum class AnimState { EDITING, COMMIT, IDLE, RESET }
 
-/**
- * Edit mode for the editor.
- */
 enum class EditMode { ADD, EDIT }
